+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * [FIX] Do not navigate to currently open profile register #244
 * [FIX] right panel should display refresh content after changes to node #249
+* [FIX] tree should be updated afgter destroy date has been set #249
 
 
 ## 3.2.5